*Version 1.7.0 +Added DPIChanged method *Enhanced SVGExplorer *Fixed issue #20: replaced Double with Single *Fixed issue #19 and 18#: Load/SaveToStream inefficient and encoding inconsistency *Fixed issue #17: Wrong conversion from pt to px *Fixed issue #14: scaling problem *Fixed issue #11: Incompatible with Drag-Drop of TImageList *Fixed issue #6: Rendering of some SVG images is incorrect

+++ GDIP/GDIPKerning.pas
@@ -44,7 +44,7 @@ type
     FLastFont: WideString;
     FLastWeight: Integer;
     FLastStyle: Integer;
-    FFontSizeFactor: Double;
+    FFontSizeFactor: Single;
 
     FFont: HFont;
     FOldFont: HFont;
@@ -171,8 +171,8 @@ end;
 procedure TGPKerningText.AddStrikeOut(const Path: TGPGraphicsPath;
   const Left, Top, Width: Single; const Font: TGPFont);
 var
-  YPos: Double;
-  Height: Double;
+  YPos: Single;
+  Height: Single;
 begin
   YPos := Top + Font.GetSize / 2;
   Height := Font.GetSize / 10;
@@ -559,8 +559,8 @@ end;
 procedure TGPKerningText.AddUnderline(const Path: TGPGraphicsPath;
   const Left, Top, Width: Single; const Font: TGPFont);
 var
-  YPos: Double;
-  Height: Double;
+  YPos: Single;
+  Height: Single;
 begin
   YPos := Top + Font.GetSize;
   Height := Font.GetSize / 10;
